requirements.txt
pytest
sqlalchemy
mysql-connector-python
pyyaml